client hold skill

+++ P1/Source/P1/Component/SkillComponent/CharacterSkillComponent.cpp
@@ -12,7 +12,7 @@ void UCharacterSkillComponent::BeginPlay()
 	SetSkills();
 }
 
-void UCharacterSkillComponent::UseSkill(uint16 SkillIndex)
+void UCharacterSkillComponent::UseSkill(int32 SkillIndex)
 {
 	Super::UseSkill(SkillIndex);
 }
