+++ P1/Source/P1/Component/SkillComponent/EnemySkillComponent.cpp
@@ -7,7 +7,7 @@ void UEnemySkillComponent::BeginPlay()
 {
 }
 
-void UEnemySkillComponent::UseSkill(uint16 SkillIndex)
+void UEnemySkillComponent::UseSkill(int32 SkillIndex)
 {
 	Super::UseSkill(SkillIndex);
 }
